+++ addons/sale_margin/data/sale_margin_demo.xml
@@ -1,6 +1,5 @@
 <?xml version="1.0" encoding="utf-8"?>
-<odoo>
-    <data noupdate="1">
+<odoo noupdate="1">
 
         <record id="sale.sale_order_line_1" model="sale.order.line">
             <field name="purchase_price">2870.00</field>
@@ -34,5 +33,4 @@
             <field name="purchase_price">13.00</field>
         </record>
 
-    </data>
 </odoo>
