+++ addons/sale_margin/models/__init__.py
@@ -2,3 +2,4 @@
 # Part of Odoo. See LICENSE file for full copyright and licensing details.
 
 from . import sale_order
+from . import sale_order_line
