+++ addons/sale_margin/views/sale_order_views.xml
